Adds versioning to corralones and testamentos
